+++ lib/flutter_gutter.dart
@@ -1,5 +1,8 @@
 export 'src/adaptive_gutter.dart';
-export 'src/axis_aware.dart';
-export 'src/breakpoints_utils.dart';
+export 'src/axis_aware_orientation.dart';
 export 'src/gutter.dart';
+export 'src/gutter_configuration.dart';
+export 'src/gutter_extensions.dart';
+export 'src/gutter_margin.dart';
+export 'src/gutter_padding.dart';
 export 'src/gutter_type.dart';
