+++ lib/src/gutter_margin.dart
@@ -1,0 +1,8 @@
+import '../flutter_gutter.dart';
+
+/// A gap of the standard margin size according to Material Design's breakpoints
+/// system
+class GutterMargin extends Gutter {
+  /// Creates a new [GutterMargin] widget.
+  const GutterMargin({super.key, super.size, super.type, super.scaleFactor});
+}
